Move task tests to separate folder
