+++ src/test/java/com/codejoust/main/task/NotificationTimerTaskTests.java
@@ -1,4 +1,4 @@
-package com.codejoust.main.util;
+package com.codejoust.main.task;
 
 import static org.junit.Assert.assertThrows;
 import static org.mockito.ArgumentMatchers.eq;
@@ -13,6 +13,7 @@ import com.codejoust.main.exception.api.ApiException;
 import com.codejoust.main.game_object.NotificationType;
 import com.codejoust.main.service.SocketService;
 
+import com.codejoust.main.util.NotificationTimerTask;
 import org.junit.Test;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
